LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src-attr-elem/strict-dynamic-elem-blocked-src-allowed.sub-expected.txt

CONSOLE MESSAGE: The source list for Content Security Policy directive 'script-src' contains an invalid source: ''strict-dynamic''. It will be ignored.
CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'script-src-elem'.

CONSOLE MESSAGE: Refused to load http://localhost:8800/content-security-policy/script-src-attr-elem/support/t_fail.js because it does not appear in the script-src directive of the Content Security Policy.

Harness Error (TIMEOUT), message = null

NOTRUN Should fire a security policy violation event
